Lights Aziz!
Looked at why there was no lighting in the ascent module cabin IVA.  Turns out I'd set the light source as an area light in Unity, which is a baked light source only, so doesn't work in KSP.  Replaced with a series of point sources to approximate an area.  Tinted red, cos it's nicer on your eyes, and makes you prepared for Duna.